move camera and view into scene folder
