+++ src/scene/scene.js
@@ -14,7 +14,7 @@ import {StyleManager} from '../styles/style_manager';
 import {Style} from '../styles/style';
 import StyleParser from '../styles/style_parser';
 import SceneLoader from './scene_loader';
-import View from '../view';
+import View from './view';
 import Light from '../light';
 import Tile from '../tile/tile';
 import TileManager from '../tile/tile_manager';
